feat: create DetailCard entity
