+++ src/entities/DetailCard/index.ts
@@ -1,0 +1,3 @@
+import DetailCard from './ui/DetailCard'
+
+export { DetailCard }
