+++ src/entities/index.ts
@@ -1,1 +1,2 @@
 export * from './CategoryCard'
+export * from './DetailCard'
